requirements.txt
streamlit
torch
torchvision
numpy
Pillow
matplotlib